rename structure to use saner names
