+++ internal/status/handlers/_testdata/nvidia/ash03.xml.faulty
