+++ internal/status/handlers/_testdata/nvidia/dump_ash01.xml
