+++ internal/status/handlers/_testdata/nvidia/dump_ash02.xml
