+++ internal/status/handlers/_testdata/nvidia/dump_ash04.xml
