+++ internal/status/handlers/_testdata/nvidia/dump_ash05.xml
