+++ internal/status/handlers/_testdata/nvidia/dump_gpu01.json
