+++ internal/status/handlers/_testdata/nvidia/dump_gpu02.json
