+++ internal/status/handlers/_testdata/nvidia/dump_gpu03.json
