+++ internal/status/handlers/_testdata/nvidia/dump_gpu04.xml
