+++ internal/status/handlers/_testdata/nvidia/dump_gpu05.xml
