+++ internal/status/handlers/_testdata/nvidia/dump_gpu10.xml
