+++ internal/status/handlers/_testdata/nvidia/dump_gpu30.xml
