+++ internal/status/handlers/_testdata/nvidia/dump_gpu36.xml
