+++ internal/status/handlers/_testdata/nvidia/dump_texel12.xml
