+++ internal/status/handlers/_testdata/nvidia/gatherdata.sh
